+++ sdk/python/lib/pulumi/output.py
@@ -40,6 +40,7 @@ U = TypeVar('U')
 
 Input = Union[T, Awaitable[T], 'Output[T]']
 Inputs = Mapping[str, Input[Any]]
+InputType = Union[T, Mapping[str, Any]]
 
 
 class Output(Generic[T]):