+++ tests/integration/types/python/declared/__main__.py
@@ -1,0 +1,90 @@
+# Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+
+from typing import Optional
+
+import pulumi
+from pulumi.dynamic import Resource, ResourceProvider, CreateResult
+
+
+@pulumi.input_type
+class AdditionalArgs:
+    def __init__(self, first_value: pulumi.Input[str], second_value: Optional[pulumi.Input[float]] = None):
+        self.first_value = first_value
+        self.second_value = second_value
+
+    # Property with empty getter/setter bodies.
+    @property
+    @pulumi.getter(name="firstValue")
+    def first_value(self) -> pulumi.Input[str]:
+        ...
+
+    @first_value.setter
+    def first_value(self, value: pulumi.Input[str]):
+        pulumi.set(self, "firstValue", value)
+
+    # Property with explicitly specified getter/setter bodies.
+    @property
+    @pulumi.getter(name="secondValue")
+    def second_value(self) -> Optional[pulumi.Input[float]]:
+        return pulumi.get(self, "secondValue")
+
+    @second_value.setter
+    def second_value(self, value: Optional[pulumi.Input[float]]):
+        pulumi.set(self, "secondValue", value)
+
+@pulumi.output_type
+class Additional(dict):
+    # Property with empty getter/setter bodies.
+    @property
+    @pulumi.getter(name="firstValue")
+    def first_value(self) -> str:
+        ...
+
+    # Property with explicitly specified getter/setter bodies.
+    @property
+    @pulumi.getter(name="secondValue")
+    def second_value(self) -> Optional[float]:
+        return pulumi.get(self, "secondValue")
+
+current_id = 0
+
+class MyResourceProvider(ResourceProvider):
+    def create(self, inputs):
+        global current_id
+        current_id += 1
+        return CreateResult(str(current_id), {"additional": inputs["additional"]})
+
+class MyResource(Resource):
+    additional: pulumi.Output[Additional]
+
+    def __init__(self, name: str, additional: pulumi.InputType[AdditionalArgs]):
+        super().__init__(MyResourceProvider(), name, {"additional": additional})
+
+
+# Create a resource with input object.
+res = MyResource("testres", additional=AdditionalArgs(first_value="hello", second_value=42))
+
+# Create a resource using the output object of another resource.
+res2 = MyResource("testres2", additional=AdditionalArgs(
+    first_value=res.additional.first_value,
+    second_value=res.additional.second_value))
+
+# Create a resource using the output object of another resource, accessing the output as a dict.
+res3 = MyResource("testres3", additional=AdditionalArgs(
+    first_value=res.additional["firstValue"],
+    second_value=res.additional["secondValue"]))
+
+# Create a resource using a dict as the input.
+res4 = MyResource("testres4", additional={
+    "firstValue": "hello",
+    "secondValue": 42,
+})
+
+pulumi.export("res_first_value", res.additional.first_value)
+pulumi.export("res_second_value", res.additional.second_value)
+pulumi.export("res2_first_value", res2.additional.first_value)
+pulumi.export("res2_second_value", res2.additional.second_value)
+pulumi.export("res3_first_value", res3.additional.first_value)
+pulumi.export("res3_second_value", res3.additional.second_value)
+pulumi.export("res4_first_value", res4.additional.first_value)
+pulumi.export("res4_second_value", res4.additional.second_value)